+++ soup/packet_writer.hpp
@@ -79,7 +79,7 @@ namespace soup
 		}
 
 		// Length-prefixed string, using u8 for the length prefix.
-		bool str_lp_u8(std::string& v, const size_t max_len = 0xFF)
+		bool str_lp_u8(std::string& v, const uint8_t max_len = 0xFF)
 		{
 			size_t len = v.size();
 			if (len <= max_len)
@@ -93,7 +93,7 @@ namespace soup
 		}
 
 		// Length-prefixed string, using u16 for the length prefix.
-		bool str_lp_u16(std::string& v, const size_t max_len = 0xFFFF)
+		bool str_lp_u16(std::string& v, const uint16_t max_len = 0xFFFF)
 		{
 			size_t len = v.size();
 			if (len <= max_len)
@@ -109,7 +109,7 @@ namespace soup
 		}
 
 		// Length-prefixed string, using u32 for the length prefix.
-		bool str_lp_u32(std::string& v, const size_t max_len = 0xFFFFFFFF)
+		bool str_lp_u32(std::string& v, const uint32_t max_len = 0xFFFFFFFF)
 		{
 			size_t len = v.size();
 			if (len <= max_len)
